+++ plugins/ci/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## 2.3.0
+## 2.4.0
 * Add Terraform scanning via a tfsec report
+
+## 2.3.0
+* Update trivy to version 0.34.0
 
 ## 2.2.4
 * Update x/text to remove CVE